Fix (most) race conditions in updating manifest
This instructs Jenkins to do a pull with rebase before attempting to update the manifest file. This should help it resolve most issues that otherwise result in the push being rejected as a non-fast-forward. Namely, if somebody updates some *other* manifest file, this will still succeed.

There is still a chance of a conflict within the prient.yml manifest file itself, in which case manual resolution is probably necessary anyways.